requirements.txt
langchain-openai
python-dotenv
requests
langchain-core
langchain-community